% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_complex_addition.R
\name{plot_complex_addition}
\alias{plot_complex_addition}
\title{Plot the sum of two complex numbers}
\usage{
plot_complex_addition(z, w)
}
\arguments{
\item{z}{A complex number.}

\item{w}{A complex number.}
}
\value{
An animated \code{ggplot} object showing the addition of the two complex numbers as vectors in the complex plane
}
\description{
This function visualizes the addition of two complex numbers \code{z} and \code{w}
as vectors in the complex plane. It plots the individual vectors and their
resulting sum \code{z + w}, illustrating complex addition geometrically.
}
\examples{
z <- 1 + 1i
w <- 2 - 1i
plot_complex_addition(z, w)

}
