% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_complex_multiplication.R
\name{plot_complex_multiplication}
\alias{plot_complex_multiplication}
\title{Plot the product of two complex numbers}
\usage{
plot_complex_multiplication(z, w, steps = 50)
}
\arguments{
\item{z}{A complex number.}

\item{w}{A complex number.}
}
\value{
An animated \code{ggplot} object showing the multiplication of the two complex numbers as vectors in the complex plane
}
\description{
This function visualizes the multiplication of two complex numbers \code{z} and \code{w}
as vectors in the complex plane. It plots the individual vectors and their
resulting product \code{z * w}, illustrating complex multiplication geometrically.
In particular, the angle (argument) is animated to show how complex multiplication
affects both magnitude and direction.
}
\examples{
z <- 2 + 1i
w <- 2 - 2i
plot_complex_multiplication(z, w)

}
