% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_unit_roots.R
\name{plot_unit_roots}
\alias{plot_unit_roots}
\title{Plot the n-th roots of unity on the unit circle}
\usage{
plot_unit_roots(n)
}
\arguments{
\item{n}{An integer (n ≥ 2). The number of complex roots of unity to display.}
}
\value{
A static \code{ggplot} object showing the roots and the corresponding polygon.
}
\description{
This function visualizes the n-th complex roots of unity by placing them
on the unit circle in the complex plane. The roots are connected in order
to form a regular n-gon.
}
\examples{
n <- 16
plot_unit_roots(n)

}
